backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==4.3.3
python-dotenv==0.21.0
google-generativeai==0.3.1
